+++ script.js
@@ -322,10 +322,15 @@ exportButton.addEventListener("click", () => {
     alert("No recent data to export.");
     return;
   }
+  let fileName = prompt("File Name:", "neovid-recent.json").trim();
+  fileName = fileName.replaceAll(" ", "-");
+  if (!fileName) {
+    fileName = "neovid-recent.json";
+  }
   const blob = new Blob([data], { type: "application/json" });
   const a = document.createElement("a");
   a.href = URL.createObjectURL(blob);
-  a.download = "neovid-recent.json";
+  a.download = fileName;
   a.click();
   URL.revokeObjectURL(a.href);
 });